requirements.txt
async-generator==1.10
attrs==22.1.0
autopep8==2.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.0.1
h11==0.14.0
idna==3.4
numpy==1.23.4
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
panda==0.3.1
pandas==1.5.1
Pillow==9.3.0
pycodestyle==2.9.1
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0
